Make the narrowing conversions explicit
All 67 bugprone/cppcoreguidelines-narrowing-conversions findings: the
implicit double->float, double->int, size_t->int and unsigned->int
conversions now go through static_cast, so the truncation is visible at
the call site instead of being silently implementation-defined.

No behaviour change; every cast reproduces the conversion the compiler
was already performing.

+++ include/ed/event_clock.h
@@ -18,7 +18,7 @@ public:
         struct timeval now{};
         gettimeofday(&now, nullptr);
 
-        double const t_secs = now.tv_sec + (now.tv_usec / 1e6);
+        double const t_secs = static_cast<double>(now.tv_sec) + (static_cast<double>(now.tv_usec) / 1e6);
         if ((t_secs - t_last_trigger_) > cycle_duration_)
         {
             t_last_trigger_ = t_secs;
